+++ Kokoro/Features/SpoofedActions.cs
@@ -14,14 +14,32 @@ partial class ApiImplementation
 	}
 }
 
-internal sealed class SpoofedActionManager
+internal sealed class SpoofedActionManager : IWrappedActionHook
 {
+	internal static readonly HiddenActionManager Instance = new();
+	
 	internal static void Setup(IHarmony harmony)
 	{
 		harmony.Patch(
 			original: AccessTools.DeclaredMethod(typeof(Card), nameof(Card.RenderAction)),
 			prefix: new HarmonyMethod(MethodBase.GetCurrentMethod()!.DeclaringType!, nameof(Card_RenderAction_Prefix))
 		);
+	}
+
+	internal static void SetupLate()
+		=> WrappedActionManager.Instance.Register(Instance, 0);
+	
+	public List<CardAction>? GetWrappedCardActions(CardAction action)
+	{
+		if (action is not ASpoofed spoofed)
+			return null;
+
+		List<CardAction> results = [];
+		if (spoofed.RenderAction is { } renderAction)
+			results.Add(renderAction);
+		if (spoofed.RealAction is { } realAction)
+			results.Add(realAction);
+		return results.Count == 0 ? null : results;
 	}
 	
 	private static bool Card_RenderAction_Prefix(G g, State state, CardAction action, bool dontDraw, int shardAvailable, int stunChargeAvailable, int bubbleJuiceAvailable, ref int __result)