+++ Kokoro/Features/WrappedActions.cs
@@ -36,15 +36,6 @@ partial class ApiImplementation
 internal sealed class WrappedActionManager : HookManager<IWrappedActionHook>
 {
 	internal static readonly WrappedActionManager Instance = new();
-	
-	public WrappedActionManager()
-	{
-		Register(ConditionalActionWrappedActionHook.Instance, 0);
-		Register(ResourceCostActionWrappedActionHook.Instance, 0);
-		Register(ContinuedActionWrappedActionHook.Instance, 0);
-		Register(HiddenActionWrappedActionHook.Instance, 0);
-		Register(SpoofedActionWrappedActionHook.Instance, 0);
-	}
 
 	internal static void Setup(IHarmony harmony)
 	{
@@ -157,88 +148,4 @@ internal sealed class WrappedActionManager : HookManager<IWrappedActionHook>
 
 	private static List<CardAction> Card_GetDataWithOverrides_Transpiler_UnwrapActions(List<CardAction> actions)
 		=> actions.SelectMany(a => Instance.GetWrappedCardActionsRecursively(a, includingWrapperActions: false)).ToList();
-}
-
-public sealed class ConditionalActionWrappedActionHook : IWrappedActionHook
-{
-	public static ConditionalActionWrappedActionHook Instance { get; private set; } = new();
-
-	private ConditionalActionWrappedActionHook() { }
-
-	public List<CardAction>? GetWrappedCardActions(CardAction action)
-	{
-		if (action is not AConditional conditional)
-			return null;
-		if (conditional.Action is not { } wrappedAction)
-			return null;
-		return [wrappedAction];
-	}
-}
-
-public sealed class ResourceCostActionWrappedActionHook : IWrappedActionHook
-{
-	public static ResourceCostActionWrappedActionHook Instance { get; private set; } = new();
-
-	private ResourceCostActionWrappedActionHook() { }
-
-	public List<CardAction>? GetWrappedCardActions(CardAction action)
-	{
-		if (action is not AResourceCost resourceCostAction)
-			return null;
-		if (resourceCostAction.Action is not { } wrappedAction)
-			return null;
-		return [wrappedAction];
-	}
-}
-
-public sealed class ContinuedActionWrappedActionHook : IWrappedActionHook
-{
-	public static ContinuedActionWrappedActionHook Instance { get; private set; } = new();
-
-	private ContinuedActionWrappedActionHook() { }
-
-	public List<CardAction>? GetWrappedCardActions(CardAction action)
-	{
-		if (action is not AContinued continued)
-			return null;
-		if (continued.Action is not { } wrappedAction)
-			return null;
-		return [wrappedAction];
-	}
-}
-
-public sealed class HiddenActionWrappedActionHook : IWrappedActionHook
-{
-	public static HiddenActionWrappedActionHook Instance { get; private set; } = new();
-
-	private HiddenActionWrappedActionHook() { }
-
-	public List<CardAction>? GetWrappedCardActions(CardAction action)
-	{
-		if (action is not AHidden hidden)
-			return null;
-		if (hidden.Action is not { } wrappedAction)
-			return null;
-		return [wrappedAction];
-	}
-}
-
-public sealed class SpoofedActionWrappedActionHook : IWrappedActionHook
-{
-	public static SpoofedActionWrappedActionHook Instance { get; private set; } = new();
-
-	private SpoofedActionWrappedActionHook() { }
-
-	public List<CardAction>? GetWrappedCardActions(CardAction action)
-	{
-		if (action is not ASpoofed spoofed)
-			return null;
-
-		List<CardAction> results = [];
-		if (spoofed.RenderAction is { } renderAction)
-			results.Add(renderAction);
-		if (spoofed.RealAction is { } realAction)
-			results.Add(realAction);
-		return results.Count == 0 ? null : results;
-	}
 }